+++ packages/builder/src/pages/builder/workspace/[application]/agent/[agentId]/knowledge/sharePointEntryTree.ts
@@ -1,0 +1,7 @@
+export interface SharePointEntryTreeNode {
+  id: string
+  name: string
+  path: string
+  type: "folder" | "file"
+  children: SharePointEntryTreeNode[]
+}
